feat: Generate video loops of AQI data
Instead of presenting live data, it's more efficient to generate images
of each hour's data and then concat them into a video loop for
presentation to the user. Here we add the components to build the
videos.

Also, some cleanup of variable names and removal of files obviated away
with the move to pre-built videos.