feat: Create videos in temp location then copy into destination
The video creation process takes up to 3 minutes for the one-year loop.
During that time, the file in the destination is unavailable. This
causes the frontend to fail. To fix this, we create the file in a temp
location and then copy it to the destination after successful creation.

Issue: CMCBMT-179